% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_utils.R
\name{as_psi_data}
\alias{as_psi_data}
\title{psiData to psi_data conversion}
\usage{
as_psi_data(psiData, parent_logger = "test")
}
\description{
This function converts an psiData to psi_data object
}
\details{
Things that this funcion do:
  List of thing to do when passing from psiData to psi_data:
  1. pl_name in plant metadata must be set to character in all sites.
  2. si_biome convert ot character (is factor)
  3. tz of solarTIMESTAMP to UTC fixed
  4. get rid of ascii characters in pl_sap_units and pl_sap_units_orig
  5. leaf data has not pl_sap_units_orig and pl_sap_units does not match
     due to bug in qc_sapw_conversion function
}
