% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_accepted_to_lvl1_psi}
\alias{df_accepted_to_lvl1_psi}
\title{Save the fixed datasets (metadata and data) in the LVL1 folder}
\usage{
df_accepted_to_lvl1_psi(
  si_code,
  psi_data = NULL,
  question_md = NULL,
  parent_logger = "test"
)
}
\arguments{
\item{si_code}{Character with the site code}

\item{psi_data}{Data frame with the fixed psi data}

\item{question_md}{Data frame with the fixed questionnaire data}
}
\value{
Nothing
}
\description{
This function performs three actions: 1) write csv files with the fixed data
in the LVL1 folder. 2) Update the status files to indicate that data is in
LVL1 and the date of the move. 3) Save the objects generated in the QC to
make easy the creation of dashboards without the need of run all QC tests
again.
}
\seealso{
Other Data Flow: 
\code{\link{df_copy_templates_psi}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_lvl1_to_lvl2}()},
\code{\link{df_lvl2_folder_structure}()},
\code{\link{df_read_psiData}()},
\code{\link{df_received_to_accepted_psi}()},
\code{\link{df_rem_to_units}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status_psi}()},
\code{\link{df_set_status_psi}()},
\code{\link{df_start_status_psi}()},
\code{\link{df_warn_to_rem}()},
\code{\link{df_whos_ready_to}()},
\code{\link{df_write_psiData}()},
\code{\link{psi_data_constructor}()}
}
\concept{Data Flow}
