% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_copy_templates_psi}
\alias{df_copy_templates_psi}
\title{Save the Rmd templates and the running scripts to their corresponding folders}
\usage{
df_copy_templates_psi(first = FALSE, parent_logger = "test")
}
\arguments{
\item{first}{Logical indicating if it is the first time the files are copied}
}
\value{
Logical indicating if all files have been copied or overwritten
}
\description{
Copy the Rmd templates for file transfer and quality check to Template folder and
the template for shiny web app to parent directory, and running scripts to parent
directory.
}
\details{
If it is the first time to set the PSI project, execute this function
after the function \code{\link{df_folder_structure}} with the argument
\code{first} set to \code{TRUE}
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1_psi}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_lvl1_to_lvl2}()},
\code{\link{df_lvl2_folder_structure}()},
\code{\link{df_read_psiData}()},
\code{\link{df_received_to_accepted_psi}()},
\code{\link{df_rem_to_units}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status_psi}()},
\code{\link{df_set_status_psi}()},
\code{\link{df_start_status_psi}()},
\code{\link{df_warn_to_rem}()},
\code{\link{df_whos_ready_to}()},
\code{\link{df_write_psiData}()},
\code{\link{psi_data_constructor}()}
}
\concept{Data Flow}
