% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_read_psiData}
\alias{df_read_psiData}
\title{Load psiData}
\usage{
df_read_psiData(
  si_code,
  level = c("Lvl_1", "Accepted"),
  parent_logger = "test"
)
}
\arguments{
\item{si_code}{Site code as a character string}

\item{level}{Level to read from as a character string}
}
\value{
A psiData object.
}
\description{
Accesory function to load an specified psiData object
}
\details{
Given a site code and a level description, \code{df_read_psiData} will return
the selected psiData object from the selected location
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1_psi}()},
\code{\link{df_copy_templates_psi}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_received_to_accepted_psi}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status_psi}()},
\code{\link{df_set_status_psi}()},
\code{\link{df_start_status_psi}()},
\code{\link{df_whos_ready_to}()},
\code{\link{psi_data_constructor}()}
}
\concept{Data Flow}
