% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_received_to_accepted_psi}
\alias{df_received_to_accepted_psi}
\title{Move from received to accepted data folders}
\usage{
df_received_to_accepted_psi(remove = FALSE, parent_logger = "test")
}
\arguments{
\item{remove}{Logical indicating if files in received folder are dropped after
the file transfer. Default to FALSE.}
}
\value{

}
\description{
Check if data file/s already exists and create the folder structure needed to
store them.
}
\details{
The first step after receiving a data set is to move it from received folder
to accepted folder (\code{Data/site_code/site_code_accepted/}).
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1_psi}()},
\code{\link{df_copy_templates_psi}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status_psi}()},
\code{\link{df_set_status_psi}()},
\code{\link{df_start_status_psi}()},
\code{\link{psi_data_constructor}()}
}
\concept{Data Flow}
