% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_rem_to_units}
\alias{df_rem_to_units}
\title{Function to pass from out_rem to out_units}
\usage{
df_rem_to_units(parent_logger = "test")
}
\value{
Nothing, all the process is internal
}
\description{
REM to UNITS transfer
}
\details{
This function is in charge of check for sites ready to pass to out_units,
perform the available transformations for that site and write the resulting
psiData in the destination folder (out_units)
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1_psi}()},
\code{\link{df_copy_templates_psi}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_lvl1_to_lvl2}()},
\code{\link{df_lvl2_folder_structure}()},
\code{\link{df_read_psiData}()},
\code{\link{df_received_to_accepted_psi}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status_psi}()},
\code{\link{df_set_status_psi}()},
\code{\link{df_start_status_psi}()},
\code{\link{df_warn_to_rem}()},
\code{\link{df_whos_ready_to}()},
\code{\link{df_write_psiData}()},
\code{\link{psi_data_constructor}()}
}
\concept{Data Flow}
