% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_warn_to_rem}
\alias{df_warn_to_rem}
\title{Function to pass from out_warn to out_rem}
\usage{
df_warn_to_rem(parent_logger = "test")
}
\value{
Nothing, all the process is internal
}
\description{
WARN to REM transfer
}
\details{
This function is in charge of check for sites ready to pass to out_rem,
remove the outliers indicated in the accesory files and saving the final
psiData objects in the corresponding folder (out_rem).
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1_psi}()},
\code{\link{df_copy_templates_psi}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_lvl1_to_lvl2}()},
\code{\link{df_lvl2_folder_structure}()},
\code{\link{df_read_psiData}()},
\code{\link{df_received_to_accepted_psi}()},
\code{\link{df_rem_to_units}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status_psi}()},
\code{\link{df_set_status_psi}()},
\code{\link{df_start_status_psi}()},
\code{\link{df_whos_ready_to}()},
\code{\link{df_write_psiData}()},
\code{\link{psi_data_constructor}()}
}
\concept{Data Flow}
