% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_write_psiData}
\alias{df_write_psiData}
\title{Write psiData}
\usage{
df_write_psiData(
  psiData,
  level = c("Lvl_1", "Lvl_2", "out_warn", "out_rem", "unit_trans"),
  units = NULL,
  parent_logger = "test"
)
}
\arguments{
\item{psiData}{psiData object}

\item{level}{Level to write to as a character string}

\item{units}{Only used if \code{level = "unit_trans"}.}
}
\value{
Nothing, the desired site data is saved.
}
\description{
Accesory function to write an psiData object to a fixed location
}
\details{
Given a site code and a level description, \code{df_write_psiData} will save
the selected psiData object in the selected location
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1_psi}()},
\code{\link{df_copy_templates_psi}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_lvl1_to_lvl2}()},
\code{\link{df_lvl2_folder_structure}()},
\code{\link{df_read_psiData}()},
\code{\link{df_received_to_accepted_psi}()},
\code{\link{df_rem_to_units}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status_psi}()},
\code{\link{df_set_status_psi}()},
\code{\link{df_start_status_psi}()},
\code{\link{df_warn_to_rem}()},
\code{\link{df_whos_ready_to}()},
\code{\link{psi_data_constructor}()}
}
\concept{Data Flow}
