% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data_loading_functions.R
\name{dl_get_si_code_psi}
\alias{dl_get_si_code_psi}
\title{Get the site code and the names of the data files}
\usage{
dl_get_si_code_psi(folder = ".", parent_logger = "test")
}
\arguments{
\item{folder}{Route to folder in which are the code and the file names to
retrieve}
}
\value{
A list. The first element is the site code, the second one the
  metadata file route, the third the psi data file route.
}
\description{
Look at the data folder provided and get the code and the names of the files
with the metadata, and psi data, in order to use them
as parameters in the automated reports
}
\seealso{
Other Data Loading Functions: 
\code{\link{dl_data_col_classes}()},
\code{\link{dl_dec_char_detect}()},
\code{\link{dl_metadata}()},
\code{\link{dl_na_char_generator}()},
\code{\link{qc_as_timestamp}()},
\code{\link{remove_dupcols}()}
}
\concept{Data Loading Functions}
