% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_psi_action}
\alias{log_psi_action}
\title{Logging handler action}
\usage{
log_psi_action(msg, handler, ...)
}
\arguments{
\item{msg}{The final formatted message to be output}

\item{handler}{The handler owning this action}
}
\description{
Custom logging handler action for psi Project
}
\seealso{
Other Logging Functions: 
\code{\link{log_psi_format}()},
\code{\link{log_psi_setup}()}
}
\concept{Logging Functions}
