% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_psi_format}
\alias{log_psi_format}
\title{Logs formatter}
\usage{
log_psi_format(record)
}
\arguments{
\item{record}{log-record object from log action}
}
\value{
Formatted message for log-record
}
\description{
Custom log formatter for psi Project
}
\seealso{
Other Logging Functions: 
\code{\link{log_psi_action}()},
\code{\link{log_psi_setup}()}
}
\concept{Logging Functions}
