% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_psi_setup}
\alias{log_psi_setup}
\title{Setup handler}
\usage{
log_psi_setup(file_name, logger, level = "DEBUG")
}
\arguments{
\item{file_name}{Name of the file in which the log will be saved}

\item{logger}{Name of the logger}

\item{level}{Level of log in character form: "FINEST", "DEBUG", "WARNING" or
"ERROR".}
}
\description{
Setting up the psi Project log handler
}
\seealso{
Other Logging Functions: 
\code{\link{log_psi_action}()},
\code{\link{log_psi_format}()}
}
\concept{Logging Functions}
