% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_work_functions.R
\name{lvl1_process}
\alias{lvl1_process}
\title{QC1 function}
\usage{
lvl1_process()
}
\description{
Function to perform level 1 data and generate reports
}
\details{
This function looks for sites without LVL1 process done and performs
level 1 QC process
}
