% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_work_functions.R
\name{lvl2_process}
\alias{lvl2_process}
\title{QC2 function, cleaning a little}
\usage{
lvl2_process(version = "0.0.1", parent_logger = "test")
}
\description{
Function to final clean the data and generate the psiQC::psi_data
objects
}
\details{
This function looks for LVL1 completed data and performs
the last cleaning and the psi_data construction. See as_psi_data for more
details
}
