% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_utils.R
\name{lvl3_process}
\alias{lvl3_process}
\title{QC3 function, cleaning a little}
\usage{
lvl3_process(version = "0.0.1", parent_logger = "test")
}
\description{
Function to final clean the data and generate the sapfluxnetr::psi_data
objects
}
\details{
This function looks for LVL2 completed data at the three levels and performs
the last cleaning and the psi_data construction. See as_psi_data for more
details
}
