% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiData_methods.R
\name{plot,psiData,missing-method}
\alias{plot,psiData,missing-method}
\title{plot psiData method}
\usage{
\S4method{plot}{psiData,missing}(x, type = c("psi", "psiSE", "psiN"), solar = FALSE)
}
\arguments{
\item{type}{what to plot}

\item{solar}{use solarTIMESTAMP?}

\item{object}{psiData object}
}
\description{
plot psiData method
}
