% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{psi_data_constructor}
\alias{psi_data_constructor}
\title{Create a PsiData object from results of Quality Checks}
\usage{
psi_data_constructor(
  psi_data = NULL,
  site_md = NULL,
  plant_md = NULL,
  question_md = NULL,
  solar_timestamp = NULL,
  parent_logger = "test"
)
}
\arguments{
\item{psi_data}{Data frame with metadata and psi data after QC process}

\item{site_md}{Data frame with site metadata after QC process}

\item{plant_md}{Data frame with plant metadata after QC process}

\item{question_md}{Data frame with questionnaire data after QC process}

\item{solar_timestamp}{Vector with solar timestamp}

\item{methods_md}{Data frame with methods metadata after QC process}
}
\value{
A PsiData object with all the data and metadata of the site
}
\description{
This function gather all the data and metadata generated in the QC process
and build an PsiData class object
}
\details{
\code{psi_data_constructor} function generates a PsiData object containing
all relevant data and metadata for a site. Add flags to psi_data as
\code{NA_ADDED}. Original NAs are also flagged as \code{NA_PRESENT}. For
info about the available slots see \code{\link{PsiData}},
\code{\link{psi_get_methods}} and \code{\link{psi_replacement}}
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1_psi}()},
\code{\link{df_copy_templates_psi}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_read_psiData}()},
\code{\link{df_received_to_accepted_psi}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status_psi}()},
\code{\link{df_set_status_psi}()},
\code{\link{df_start_status_psi}()},
\code{\link{df_whos_ready_to}()}
}
\concept{Data Flow}
