% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiData_generics.R
\name{psi_get_generics}
\alias{psi_get_generics}
\alias{get_psi}
\alias{get_psi_flags}
\alias{get_timestamp}
\alias{get_solar_timestamp}
\alias{get_si_code}
\alias{get_site_md}
\alias{get_plant_md}
\alias{get_question_md}
\title{psiData custom get generics}
\usage{
get_psi(object, ...)

get_psi_flags(object, ...)

get_timestamp(object, ...)

get_solar_timestamp(object, ...)

get_si_code(object, ...)

get_site_md(object, ...)

get_plant_md(object, ...)

get_question_md(object, ...)
}
\description{
Generics for getting the info in the slots of psiData
}
