% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiData_methods.R
\name{psi_get_methods}
\alias{psi_get_methods}
\alias{get_psi,psiData-method}
\alias{get_psi_flags,psiData-method}
\alias{get_timestamp,psiData-method}
\alias{get_solar_timestamp,psiData-method}
\alias{get_si_code,psiData-method}
\alias{get_site_md,psiData-method}
\alias{get_question_md,psiData-method}
\alias{get_plant_md,psiData-method}
\title{psiData get methods}
\usage{
\S4method{get_psi}{psiData}(object, solar = FALSE)

\S4method{get_psi_flags}{psiData}(object, solar = FALSE)

\S4method{get_timestamp}{psiData}(object)

\S4method{get_solar_timestamp}{psiData}(object)

\S4method{get_si_code}{psiData}(object)

\S4method{get_site_md}{psiData}(object)

\S4method{get_question_md}{psiData}(object)

\S4method{get_plant_md}{psiData}(object)
}
\arguments{
\item{object}{Object of class psiData from which data is retrieved}

\item{solar}{Logical indicating if the timestamp to return in the get_psi
and get_psi_flags methods}
}
\description{
Methods to get the info from the psiData class slots
}
\details{
\code{get_psi} method retrieve psi data and timestamp to create a functional
dataset to work with.

\code{get_psi_flags} method retrieve sapflow or environmental flags also
with the timestamp.

\code{get_timestamp} method retrieve only the timestamp as POSIXct vector.

\code{get_si_code} method retrieve a character vector with length(timestamp)
containing the site code.

\code{get_site_md}, and \code{get_plant_md} methods retrieve the corresponding
metadata.
}
