% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiData_methods.R
\name{psi_replacement}
\alias{psi_replacement}
\alias{get_psi<-,psiData-method}
\alias{get_psi_flags<-,psiData-method}
\alias{get_timestamp<-,psiData-method}
\alias{get_solar_timestamp<-,psiData-method}
\alias{get_si_code<-,psiData-method}
\alias{get_site_md<-,psiData-method}
\alias{get_plant_md<-,psiData-method}
\title{Replacement methods}
\usage{
\S4method{get_psi}{psiData}(object) <- value

\S4method{get_psi_flags}{psiData}(object) <- value

\S4method{get_timestamp}{psiData}(object) <- value

\S4method{get_solar_timestamp}{psiData}(object) <- value

\S4method{get_si_code}{psiData}(object) <- value

\S4method{get_site_md}{psiData}(object) <- value

\S4method{get_plant_md}{psiData}(object) <- value
}
\value{
The same psiData object with the corresponding slot changed to the
  value provided. An error if the value provided generates an invalid
  psiData object.
}
\description{
Methods for replacing the slots with new data or metadata
}
\details{
The replacement object must be a valid object for that slot, i.e. for psilow
data slot a data frame with the same dimensions and without TIMESTAMP variable
is needed. A validity check is done before returning the replaced psiData
object and an error is returned if this check fails.
}
