% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiData_generics.R
\name{psi_replacement_generics}
\alias{psi_replacement_generics}
\alias{get_psi<-}
\alias{get_psi_flags<-}
\alias{get_timestamp<-}
\alias{get_solar_timestamp<-}
\alias{get_si_code<-}
\alias{get_site_md<-}
\alias{get_plant_md<-}
\alias{get_question_md<-}
\title{Replacement generics}
\usage{
get_psi(object) <- value

get_psi_flags(object) <- value

get_timestamp(object) <- value

get_solar_timestamp(object) <- value

get_si_code(object) <- value

get_site_md(object) <- value

get_plant_md(object) <- value

get_question_md(object) <- value
}
\description{
Generic functions for replacement functions for psiData
}
