% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_work_functions.R
\name{psi_tidyfier}
\alias{psi_tidyfier}
\title{PSI tidyfier}
\usage{
psi_tidyfier(
  psi_data,
  include_flags = FALSE,
  include_question = FALSE,
  parent_logger = "test"
)
}
\arguments{
\item{psi_data}{A psiData object}

\item{include_flags}{A logical object. Set to TRUE if data flags should be
included in the returned object}

\item{include_question}{A logical object. Set to TRUE if data sets questionnaire
data should be included in the returned object}
}
\value{
tibble object
}
\description{
Function to transform psiData objects to tidy format
}
