% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outliers_and_ranges.R
\name{qc_env_ranges}
\alias{qc_env_ranges}
\title{Check ranges for environmental variables}
\usage{
qc_env_ranges(env_data, env_flags, parent_logger = "test")
}
\arguments{
\item{env_data}{Environmental data, as obtained from \code{\link{get_env}}}

\item{env_flags}{Environmental flags, as obtained from \code{\link{get_env_flags}}}
}
\value{
A data frame containing the new flags in case any value is out of
  range.
}
\description{
Check ranges for environmental variables
}
\details{
Environmental values are checked against known general ranges
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_timezone_coord}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_psi_N_nas}()},
\code{\link{qc_psi_SE_nas}()},
\code{\link{qc_psi_nas}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_simplify_questions}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process_psi}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
