% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_get_timezone_coord}
\alias{qc_get_timezone_coord}
\title{Timezones from coordinates}
\usage{
qc_get_timezone_coord(coord_lat, coord_lon, parent_logger = "test")
}
\arguments{
\item{coord_lat}{numeric vector of latitudes}

\item{coord_lon}{numeric vector of longitudes the same length as x}
}
\value{
A character vector with the timezone code compatible with lubridate
  and as.POSIXct
}
\description{
Get timezones from coordinates using lutz package and tranforms timezone ISO
code to character vector compatible with lubridate and POSIXct
}
\details{
GMT time zones are used, as they are day saving light time (DST) agnostic,
and in that way the DST can setted if the metadata says so. GMT are sign
exchanged to be compatible with ISO.
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_psi_N_nas}()},
\code{\link{qc_psi_SE_nas}()},
\code{\link{qc_psi_nas}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_simplify_questions}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process_psi}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
