% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_is_timestamp}
\alias{qc_is_timestamp}
\title{TimeStamp format check}
\usage{
qc_is_timestamp(data, verbose = TRUE, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the data (sapflow or environmental). Also
a vector with the timestamp values. If data frame, \bold{it must contain} a
timestamp variable}

\item{verbose}{Logical indicating if messages of success and warnings of
failures must be presented. Default is \code{TRUE}. In order to use
inside another function, is recommended to set \code{verbose = FALSE}.}
}
\value{
A message/warning indicating if timestamp is correct or not. Also
  an invisible logical object is returned, indicating success (TRUE) or
  failure (FALSE).
}
\description{
Function to check if the format of the timestamp is the correct one
}
\details{
Template timestamp is requiered to comply POSIXct format. This function
checks if the contributor followed the mandatory format.
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_timezone_coord}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_psi_N_nas}()},
\code{\link{qc_psi_SE_nas}()},
\code{\link{qc_psi_nas}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_simplify_questions}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process_psi}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
