% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{qc_md_cols}
\alias{qc_md_cols}
\title{Metadata columns check}
\usage{
qc_md_cols(metadata, dic, parent_logger = "test")
}
\arguments{
\item{metadata}{Data frame containing the data or metadata in which the test
will be made.}

\item{dic}{Name of the metadata dictionary to use as character. It must be
one of the following: \code{'site_md'}, \code{'plant_md'},
\code{'psi_data'}, or \code{'Questionnaire'}}
}
\value{
A data frame with variable names in one column and result of the
  checks in the following columns
}
\description{
\code{qc_md_cols} checks if the columns of the provided metadata return the
correct class, as well as indicates any NAs present and the absence of any
mandatory variable.
}
\details{
After loading the metadata, columns classes and presence/absence of mandatory
variables must be checked in order to continue with the quality check
workflow. This function returns a summary of the metadata columns and their
state.
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_timezone_coord}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_simplify_questions}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process_psi}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
