% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{qc_measured_sfn}
\alias{qc_measured_sfn}
\title{Correct measured_sfn values}
\usage{
qc_measured_sfn(plant_md, parent_logger = "test")
}
\arguments{
\item{plant_md}{Data frame containing the plant metadata}
}
\value{
A data frame with the information about the different treatments if
  any.
}
\description{
\code{qc_measured_sfn} function standardizes the values provided by the authors.
}
\details{
This function standardizes the values provided by the authors which could be
\code{yes}, \code{YES}, \code{no}, \code{NO}, \code{TRUE}, \code{FALSE}
\code{Yes} or \code{No}.
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_timezone_coord}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_psi_N_nas}()},
\code{\link{qc_psi_SE_nas}()},
\code{\link{qc_psi_nas}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_simplify_questions}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process_psi}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
