% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_set_timezone}
\alias{qc_set_timezone}
\title{Set the timezone of the timestamp}
\usage{
qc_set_timezone(data, tz, parent_logger = "test")
}
\arguments{
\item{data}{Data frame with the timestamp variable to set, or a POSIXct
vector.}

\item{tz}{Character vector with the compatible name of the timezone, as the
one provided by the \code{\link{qc_get_timezone}} function.}
}
\value{
A data frame as the \code{data} provided, with the timestamp variable
  associated to the timezone specified
}
\description{
Brute force convert of timezone
}
\details{
When reading data from xlsx or csv, timestamp is readed as POSIXct and by
default the timezone is UTC. With this function timezone can be changed
without change the timestamp. This is made with the \code{force_tz} function
of the lubridate package.
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_timezone_coord}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_psi_N_nas}()},
\code{\link{qc_psi_SE_nas}()},
\code{\link{qc_psi_nas}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_simplify_questions}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process_psi}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
