% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_names}
\alias{qc_species_names}
\title{Wrapper for species names check}
\usage{
qc_species_names(species, max_distance = 1, parent_logger = "test")
}
\arguments{
\item{species}{Character vector containing the species names to verify}

\item{max_distance}{NA number indicating the maximum distance allowed for
a match in agrep when performing corrections of spelling errors in
specific epithets. Not used if corr = FALSE.}
}
\value{
a vector with the fixed names of the species if fix is possible or
  needed, or a vector with the original names of the species if the fix is not
  possible or they are correct
}
\description{
Wrapper for \code{\link{qc_species_names_info}} and \code{\link{qc_species_names_fix}}
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_timezone_coord}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_measured_sfn}()},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_psi_N_nas}()},
\code{\link{qc_psi_SE_nas}()},
\code{\link{qc_psi_nas}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_simplify_questions}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process_psi}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
