% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_names_fix}
\alias{qc_species_names_fix}
\title{Verify provided species names (spelling and correctness)}
\usage{
qc_species_names_fix(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame as the obtained from \code{\link{qc_species_names_info}}}
}
\value{
A character vector with species fixed in spelling and correctness.
}
\description{
\code{qc_species_names} uses \code{Taxonstand} package.
}
\details{
This function takes a vector of species names and check if they are right
spelled. Also, if a synonym is used, the function changes it automatically in
order to have the same name for the same species
}
