% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_names_info}
\alias{qc_species_names_info}
\title{Info of species names spelling}
\usage{
qc_species_names_info(species, max_distance = 1, parent_logger = "test")
}
\arguments{
\item{species}{Character vector containing the species names to verify}

\item{max_distance}{A number indicating the maximum distance allowed
for a match in agrep}
}
\value{
A data frame summarizing the species names declared, the species
  names obtained after tpl and the concordance and NAs info
}
\description{
Summary of species names spelling info
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_timezone_coord}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_measured_sfn}()},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_psi_N_nas}()},
\code{\link{qc_psi_SE_nas}()},
\code{\link{qc_psi_nas}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_simplify_questions}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process_psi}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
