% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_work_functions.R
\name{read_psi_data}
\alias{read_psi_data}
\title{Read psi_data from disk}
\usage{
read_psi_data(site_code, folder = ".")
}
\arguments{
\item{site_code}{A character vector with one site code}

\item{folder}{Route to the folder containing the \code{.RData} file. Default
to working directory.}
}
\value{
an psi_data object with the selected site data.
}
\description{
Given a site code and a route, \code{read_psi_data} will return the selected
psi_data object
}
\examples{
# Let's access the data in "folder". This typically is the folder where the
# psi data is, but in this example we will create a temporal folder with some
# data to test the function
folder <- tempdir()
save(ARG_TRE, file = file.path(folder, 'ARG_TRE.RData'))
save(ARG_MAZ, file = file.path(folder, 'ARG_MAZ.RData'))

# now we read a single site
ARG_TRE_test <- read_psi_data('ARG_TRE', folder)
ARG_TRE_test


}
