% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiData_methods.R
\name{show,psiData-method}
\alias{show,psiData-method}
\title{Show method for psiData}
\usage{
\S4method{show}{psiData}(object)
}
\arguments{
\item{object}{psiData object to show}
}
\description{
Show method for psiData
}
