% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiData_methods.R
\name{[,psiData,numeric,ANY,missing-method}
\alias{[,psiData,numeric,ANY,missing-method}
\title{Sub-setting operation}
\usage{
\S4method{[}{psiData,numeric,ANY,missing}(x, i, j)
}
\arguments{
\item{i}{data row index}

\item{j}{psi data column index}

\item{object}{psiData object}
}
\description{
Sub-setting operation
}
