% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{write_psi_data}
\alias{write_psi_data}
\title{Function to save the psi_data created}
\usage{
write_psi_data(psi_data, folder, parent_logger = "test")
}
\description{
This function saves the RData files with the psi_data objects
}
