% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{psi_data2csv}
\alias{psi_data2csv}
\title{psi_data2csv function}
\usage{
psi_data2csv(psi_data, csv_folder)
}
\description{
This function is used in lvl2_process to write the csv files for each psi_data object
slots in the corresponding folder of the database
}
